requirements.txt
mlflow==2.20.2
ipykernel
python-dotenv 
